New & Improved client interface!
This rewrite came with the currently-ongoing architectural changes in
teahaz-server. The overarching theme is to move all server logic to a
`Chatroom` object, and let the main object only deal with organized
interactions to it.

In practice, this means that:
1. `Bot()` is now `Teacup()`, but with a lot of differences
2. sending, receiving & manipulating data is now done through
   `Chatroom`, which can be acquired through various means from the
   `Teacup()`
3. There is a much better event system, which now only requires a single
   handler function for each purpose, thus making a simple bot has gone
   from a full class override to a simple function definition & a call
   to `Chatroom().subscribe`.

The code in general is much neater too: pylint now reports 10.00/10 for
its quality.

Most things are currently not tested due to some server issues, but they
are built following the spec, so they should be aight.

Important: This interface does not currently support all server methods,
but it should be feature-complete shortly.