Added support for `/messages/` `GET` by time!
This is a really important feature in terms of client development, not
so much the end user. There are some weird hoops the program has to jump
through (duplicate filtering for the most part) but it works well.

There is also now a nice test for message sending, which sends a
variable amount of messages and looks if it can get all of them back.